feat: integrate DB with server

+++ src/main/java/james/expense_tracker/dto/expense/CreateExpenseRequest.java
@@ -1,5 +1,7 @@
 package james.expense_tracker.dto.expense;
 
+import java.math.BigDecimal;
+
 import james.expense_tracker.model.ExpenseType;
 
-public record CreateExpenseRequest(String description, double amount, ExpenseType type) {}
+public record CreateExpenseRequest(String description, BigDecimal amount, ExpenseType type) {}
